Fix for AttributeError in non-blocking example
Since s.open_session() takes some time, it might happen that chan has the value of LIBSSH2_ERROR_EAGAIN. This is not handled in the code. I've added an other loop that checks for None, which is now the initial value of chan, and LIBSSH2_ERROR_EAGAIN and tries to open the session again in this cases.